feat(node): Node client extends `ServerRuntimeClient` rather than `BaseClient` (#8933)
Since common server bahaviour has now moved to `ServerRuntimeClient`,
the Node client can now extend that rather than the base client.

This PR also moves the request session flusher to `ServerRuntimeClient`
which leaves the Node client empty apart from a constructor wrapper to
preserve backwards compatibilty.
